requirements.txt
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
